+++ tests/back_compat/test_read.py
@@ -43,7 +43,7 @@ class TestReadOldVersions(TestCase):
             with self.subTest(file=f.name):
                 with warnings.catch_warnings(record=True) as warnings_on_read:
                     warnings.simplefilter("always")
-                    with NWBHDF5IO(str(f), 'r') as io:
+                    with NWBHDF5IO(str(f), 'r', load_namespaces=True) as io:
                         errors = validate(io)
                         io.read()
                         for w in warnings_on_read: